fix UI build + eslint + DB migration timestamps + PR comments
