+++ db/structure.sql
@@ -3991,7 +3991,7 @@ SET row_security = off;
 --
 
 COPY public.schema_migrations (version, dirty) FROM stdin;
-20260901071557	f
+20260901143615	f
 \.
