+++ pkg/components/runner/runner_task_lifecycle.go
@@ -128,9 +128,8 @@ func processBrokerTaskStatus(
 		return fmt.Errorf("task is not in terminal state")
 	}
 
-	// Persist spend even when SuperPlane already cancelled the node.
-	// Check-fix runs are cancelled when a new revision arrives; the broker
-	// webhook still carries the tokens the agent already billed.
+	// Persist spend when the broker reports a terminal task after SuperPlane
+	// already finished the node. A late webhook still carries billed tokens.
 	publishRunnerUsage(organizationID, task, logger)
 	RecordRunnerLLMUsage(usage, logger, finishedEventType, configuration, task.Result)
 	RecordRunnerComputeUsage(usage, logger, state, configuration, task)
@@ -192,10 +191,10 @@ func billableSeconds(duration time.Duration) int64 {
 }
 
 func cancelBrokerTask(ctx core.ExecutionContext) error {
-	return cancelBrokerTaskWithUsage(ctx, "")
-}
+	if ctx.ExecutionState.IsFinished() {
+		return nil
+	}
 
-func cancelBrokerTaskWithUsage(ctx core.ExecutionContext, finishedEventType string) error {
 	taskID, err := ctx.ExecutionState.GetKV("task_id")
 	if err != nil {
 		if errors.Is(err, core.ErrExecutionKVNotFound) {
@@ -209,27 +208,8 @@ func cancelBrokerTaskWithUsage(ctx core.ExecutionContext, finishedEventType stri
 		return err
 	}
 
-	if finishedEventType != "" {
-		recordUsageFromBrokerTask(ctx, broker, taskID, finishedEventType)
-	}
-
-	if ctx.ExecutionState.IsFinished() {
-		return nil
-	}
-
 	if err := broker.CancelTask(taskID); err != nil {
 		return fmt.Errorf("cancel task: %w", err)
 	}
 	return nil
-}
-
-func recordUsageFromBrokerTask(ctx core.ExecutionContext, broker *BrokerClient, taskID, finishedEventType string) {
-	task, err := broker.FetchTaskStatus(taskID)
-	if err != nil {
-		if ctx.Logger != nil {
-			ctx.Logger.WithError(err).Warn("runner: failed to fetch task status before cancel")
-		}
-		return
-	}
-	RecordRunnerLLMUsage(ctx.Usage, ctx.Logger, finishedEventType, ctx.Configuration, task.Result)
 }
